Move compression tests to new file tests/io/test_compression.py
